final changes for pmax 4.4-Lite II release
SCCS-vsn: sys/pmax/include/param.h 8.4
SCCS-vsn: sys/pmax/include/param.h 8.5
SCCS-vsn: sys/pmax/include/types.h 8.4
SCCS-vsn: sys/pmax/pmax/conf.c 8.4
SCCS-vsn: sys/pmax/pmax/machdep.c 8.5
SCCS-vsn: sys/pmax/pmax/genassym.c 8.3
SCCS-vsn: sys/pmax/pmax/mem.c 8.3
SCCS-vsn: sys/pmax/pmax/pmap.c 8.5
SCCS-vsn: sys/pmax/pmax/trap.c 8.7
SCCS-vsn: sys/pmax/pmax/locore.s 8.7
SCCS-vsn: sys/pmax/pmax/pmaxtype.h 8.2
SCCS-vsn: sys/pmax/dev/dc.c 8.5
SCCS-vsn: sys/pmax/dev/pm.c 8.2
SCCS-vsn: sys/pmax/dev/rz.c 8.4
SCCS-vsn: sys/pmax/dev/tz.c 8.5
SCCS-vsn: sys/pmax/dev/cfb.c 8.2
SCCS-vsn: sys/pmax/dev/mfb.c 8.2
SCCS-vsn: sys/pmax/dev/scc.c 8.3
SCCS-vsn: sys/pmax/dev/dtop.c 8.3
SCCS-vsn: sys/pmax/dev/xcfb.c 8.2

+++ usr/src/sys/pmax/dev/scc.c
@@ -7,7 +7,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)scc.c	8.2 (Berkeley) %G%
+ *	@(#)scc.c	8.3 (Berkeley) %G%
  */
 
 /* 
@@ -394,7 +394,7 @@ sccwrite(dev, uio, flag)
 /*ARGSUSED*/
 sccioctl(dev, cmd, data, flag, p)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	int flag;
 	struct proc *p;
